made varients comptible with linux builds.
As the anaconda compiler is set to use C++17 standard on linux (differs
from the C++14 standard on mac) the language changes are note yet supported
by ITK. In addition, the VMTK source code has not been updated to support these
language changes. Unlike the C++14 standard used on macs, the C++17 standards
throw fatal errors at compile time (not just warnings).

It is still best to compile with C++11 on linux. It is worth using C++14 on mac
(with the anaconda compilers) in order to take advantage of the build system using
the MacOSX10.9 SDKs, which make the system compatible with all 10.9+ deployment targets.